# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Calculator Service","description":"Service for calc numbers","version":"1.0"},"servers":[{"url":"http://localhost:8000"}],"paths":{"/div/{a}/{b}":{"get":{"tags":["calc"],"summary":"div calc","operationId":"calc#div","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":207017106558923967,"format":"int64"},"example":6001261670528037577},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":5850431520333673251,"format":"int64"},"example":1089204046671954241}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":1016925590796087323,"format":"int64"},"example":3833729960400405184}}},"400":{"description":"DivByZero: Bad Request response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/multiply/{a}/{b}":{"get":{"tags":["calc"],"summary":"multiply calc","operationId":"calc#multiply","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":4886963557863946648,"format":"int64"},"example":6642806206162570878},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":8723986392256123438,"format":"int64"},"example":9092197569596372935}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":5308928064408640979,"format":"int64"},"example":2937026320033011362}}}}}},"/openapi.json":{"get":{"tags":["calc"],"summary":"Download ./gen/http/openapi3.json","operationId":"calc#/openapi.json","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"},"required":["name","id","message","temporary","timeout","fault"]}}},"tags":[{"name":"calc","description":"The calc service performs operations on numbers."}]}